Refactor: Rename vectors to vector_db
